Extract diagnostics emitting into separate files
