+++ crates/mockiato-codegen/src/lib.rs
@@ -19,6 +19,7 @@ extern crate proc_macro;
 
 mod constant;
 mod diagnostic;
+mod emit_diagnostics;
 mod generate;
 mod mockable;
 mod parse;
@@ -26,20 +27,9 @@ mod result;
 mod syn_ext;
 
 use self::mockable::Mockable;
-#[cfg(rustc_is_nightly)]
-use crate::diagnostic::{Diagnostic, DiagnosticLevel, DiagnosticMessage};
-use crate::result::Error;
+use crate::emit_diagnostics::emit_diagnostics;
 use proc_macro::TokenStream as ProcMacroTokenStream;
-#[cfg(rustc_is_nightly)]
-use proc_macro::{
-    Diagnostic as ProcMacroDiagnostic, Level as ProcMacroLevel, Span as ProcMacroSpan,
-};
-#[cfg(rustc_is_nightly)]
-use proc_macro2::Span;
 use syn::{parse_macro_input, AttributeArgs, Item};
-use proc_macro2::TokenStream;
-#[cfg(not(rustc_is_nightly))]
-use quote::quote_spanned;
 
 #[doc(hidden)]
 #[proc_macro_attribute]
@@ -61,73 +51,5 @@ pub fn mockable(args: ProcMacroTokenStream, input: ProcMacroTokenStream) -> Proc
 
             output
         }
-    }
-}
-
-#[cfg(not(rustc_is_nightly))]
-fn emit_diagnostics(error: Error) -> TokenStream {
-    error
-        .diagnostics
-        .into_iter()
-        .map(|diagnostic| {
-            let message = diagnostic.message;
-            quote_spanned!(diagnostic.span => compile_error!(#message);)
-        })
-        .collect()
-}
-
-#[cfg(rustc_is_nightly)]
-fn emit_diagnostics(error: Error) -> TokenStream {
-    error
-        .diagnostics
-        .into_iter()
-        .map(to_proc_macro_diagnostic)
-        .for_each(ProcMacroDiagnostic::emit);
-    TokenStream::new()
-}
-
-#[cfg(rustc_is_nightly)]
-fn to_proc_macro_diagnostic(source: Diagnostic) -> ProcMacroDiagnostic {
-    let level = to_proc_macro_level(source.level);
-    let span = to_proc_macro_span(source.span);
-    let diagnostic = ProcMacroDiagnostic::spanned(span, level, source.message);
-    let diagnostic = add_notes_to_proc_macro_diagnostic(diagnostic, source.notes);
-    add_help_to_proc_macro_diagnostic(diagnostic, source.help)
-}
-
-#[cfg(rustc_is_nightly)]
-fn add_help_to_proc_macro_diagnostic(
-    diagnostic: ProcMacroDiagnostic,
-    help: Vec<DiagnosticMessage>,
-) -> ProcMacroDiagnostic {
-    help.into_iter()
-        .fold(diagnostic, |diagnostic, help| match help.span {
-            Some(span) => diagnostic.span_help(to_proc_macro_span(span), help.message),
-            None => diagnostic.help(help.message),
-        })
-}
-
-#[cfg(rustc_is_nightly)]
-fn add_notes_to_proc_macro_diagnostic(
-    diagnostic: ProcMacroDiagnostic,
-    notes: Vec<DiagnosticMessage>,
-) -> ProcMacroDiagnostic {
-    notes
-        .into_iter()
-        .fold(diagnostic, |diagnostic, note| match note.span {
-            Some(span) => diagnostic.span_note(to_proc_macro_span(span), note.message),
-            None => diagnostic.note(note.message),
-        })
-}
-
-#[cfg(rustc_is_nightly)]
-fn to_proc_macro_span(span: Span) -> ProcMacroSpan {
-    span.unstable()
-}
-
-#[cfg(rustc_is_nightly)]
-fn to_proc_macro_level(level: DiagnosticLevel) -> ProcMacroLevel {
-    match level {
-        DiagnosticLevel::Error => ProcMacroLevel::Error,
     }
 }
